tblcheck: Grade Tables in Learning Exercises

'tblcheck' extends 'gradethis' with functions that inspect data frame or tibble objects to make it easier for teachers to check that student tables meet expectations.

See also

Author

Maintainer: Alexander Rossell Hayes alex.rossellhayes@rstudio.com (ORCID)

Authors:

Garrick Aden-Buie garrick@rstudio.com (ORCID)

Other contributors:

Sara Altman sara.altman@rstudio.com (ORCID) [contributor]

RStudio, PBC [copyright holder, funder]